impl vertical scroll

+++ src/config.rs
@@ -1,8 +1,8 @@
 use termion::terminal_size;
 
 pub struct EditorCfg<'a> {
-    pub screen_row: u16,
-    pub screen_col: u16,
+    pub screen_row: u32,
+    pub screen_col: u32,
     pub file_name: &'a str,
 }
 
@@ -10,8 +10,8 @@ impl<'a> EditorCfg<'a> {
     pub fn new(file_name: &'a str) -> Self {
         let size = terminal_size().unwrap();
         EditorCfg {
-            screen_col: size.0,
-            screen_row: size.1,
+            screen_col: size.0 as u32,
+            screen_row: size.1 as u32,
             file_name,
         }
     }
